workers/requirements.txt
celery==5.4.0
redis==5.2.1
requests==2.32.3
pymongo==4.10.1
python-dotenv==1.0.1
